+++ common/src/main/java/net/more_rpg_classes/client/particle/MoreParticles.java
@@ -33,6 +33,7 @@ public class MoreParticles {
     public static final SimpleParticleType ICE_TRAP = FabricParticleTypes.simple();
     public static final SimpleParticleType STONE_TRAP = FabricParticleTypes.simple();
     public static final SimpleParticleType LEAF = FabricParticleTypes.simple();
+    public static final SimpleParticleType FATAL_POISON = FabricParticleTypes.simple();
     public static final SimpleParticleType RAGE_PAR = FabricParticleTypes.simple();
     public static final SimpleParticleType SMALL_THUNDER = FabricParticleTypes.simple();
 
@@ -63,6 +64,7 @@ public class MoreParticles {
         Registry.register(Registries.PARTICLE_TYPE, Identifier.of(MRPGCMod.MOD_ID, "ice_trap"), ICE_TRAP);
         Registry.register(Registries.PARTICLE_TYPE, Identifier.of(MRPGCMod.MOD_ID, "stone_trap"), STONE_TRAP);
         Registry.register(Registries.PARTICLE_TYPE, Identifier.of(MRPGCMod.MOD_ID, "leaf"), LEAF);
+        Registry.register(Registries.PARTICLE_TYPE, Identifier.of(MRPGCMod.MOD_ID, "fatal_poison"), FATAL_POISON);
         Registry.register(Registries.PARTICLE_TYPE, Identifier.of(MRPGCMod.MOD_ID, "rage_particle"), RAGE_PAR);
         Registry.register(Registries.PARTICLE_TYPE, Identifier.of(MRPGCMod.MOD_ID, "small_thunder"), SMALL_THUNDER);
     }